extract fixtures
